chg: dev: make the autologin feature work on every page (fixes #82)
