+++ src/components/Core.vue
@@ -19,26 +19,10 @@ import { Options, Vue } from 'vue-class-component';
 import LeftCol from "./core/LeftCol.vue";
 import RightCol from "./core/RightCol.vue";
 import SendAskMoney from "./sendAskMoney/SendAskMoney.vue";
-import router from "../router/index"
 
 
 @Options({
   name:"core",
-  async mounted() {
-    // If this component is loaded, we already are logged in, so
-    // the store already contains the logged in user informations...
-    // Maybe all the following is not useful ?
-    if(!this.userProfile) {
-      try {
-        await this.$store.dispatch("initAutoLogin")
-        await this.$store.dispatch("setAccounts")
-      } catch(e) {
-        console.error("Error while trying to autolog", e)
-        router.push("/")
-        throw e
-      }
-    }
-  },
   computed: {
     // We do not have to rely on an extra db call to get the
     // information, just use a computed property on the store state (or getter if needed)
